test(rsc): migrate hoist output fixtures
