+++ packages/plugin-rsc/src/transforms/fixtures/hoist/arrow.js
@@ -1,0 +1,13 @@
+let count = 0
+
+function Counter() {
+  const name = 'value'
+
+  return {
+    type: 'form',
+    action: (formData) => {
+      'use server'
+      count += Number(formData.get(name))
+    },
+  }
+}
