+++ packages/plugin-rsc/src/transforms/fixtures/hoist/arrow.js.snap.encode.js
@@ -1,0 +1,17 @@
+let count = 0
+
+function Counter() {
+  const name = 'value'
+
+  return {
+    type: 'form',
+    action: /* #__PURE__ */ $$register($$hoist_0_anonymous_server_function, "<id>", "$$hoist_0_anonymous_server_function").bind(null, __enc([name])),
+  }
+}
+
+;export function $$hoist_0_anonymous_server_function($$hoist_encoded, formData) {
+      const [name] = __dec($$hoist_encoded);
+'use server'
+      count += Number(formData.get(name))
+    };
+/* #__PURE__ */ Object.defineProperty($$hoist_0_anonymous_server_function, "name", { value: "anonymous_server_function" });
