+++ packages/plugin-rsc/src/transforms/fixtures/hoist/closure.js
@@ -1,0 +1,12 @@
+let count = 0
+
+function Counter() {
+  const name = 'value'
+
+  async function changeCount(formData) {
+    'use server'
+    count += Number(formData.get(name))
+  }
+
+  return 'something'
+}
