+++ packages/plugin-rsc/src/transforms/fixtures/hoist/multiple.js
@@ -1,0 +1,17 @@
+let count = 0
+
+function Counter() {
+  const name = 'value'
+
+  async function changeCount(formData) {
+    'use server'
+    count += Number(formData.get(name))
+  }
+
+  async function changeCount2(formData) {
+    'use server'
+    count += Number(formData.get(name))
+  }
+
+  return 'something'
+}
