+++ packages/plugin-rsc/src/transforms/hoist.test.ts
@@ -125,7 +125,7 @@ async function f() {
   return x;
 }
 `
-    expect(await testTransform(input)).toMatchInlineSnapshot(`undefined`)
+    expect(await testTransform(input)).toBeUndefined()
   })
 
   it('ignores strings outside a function directive prologue', async () => {
@@ -152,6 +152,21 @@ async function action() {
     expect(await testTransformNames(input)).toEqual(['$$hoist_0_action'])
   })
 
+  it('returns generated names in order', async () => {
+    const input = `
+async function first() {
+  "use server";
+}
+async function second() {
+  "use server";
+}
+`
+    expect(await testTransformNames(input)).toEqual([
+      '$$hoist_0_first',
+      '$$hoist_1_second',
+    ])
+  })
+
   it('finds directives only in directive-capable bodies', async () => {
     const input = `
 {
@@ -163,322 +178,6 @@ async function action() {
 `
     const ast = await parseAstAsync(input)
     expect(findDirectives(ast, 'use server')).toHaveLength(1)
-  })
-
-  it('top level', async () => {
-    const input = `
-const x = "x";
-
-async function f() {
-  "use server";
-  return x;
-}
-
-async function g() {
-}
-
-export async function h(formData) {
-  "use server";
-  return formData.get(x);
-}
-
-export default function w() {
-  "use server";
-}
-`
-    expect(await testTransform(input)).toMatchInlineSnapshot(`
-      "
-      const x = "x";
-
-      const f = /* #__PURE__ */ $$register($$hoist_0_f, "<id>", "$$hoist_0_f");
-
-      async function g() {
-      }
-
-      export const h = /* #__PURE__ */ $$register($$hoist_1_h, "<id>", "$$hoist_1_h");
-
-      const w = /* #__PURE__ */ $$register($$hoist_2_w, "<id>", "$$hoist_2_w");
-      export default w;
-
-      ;export async function $$hoist_0_f() {
-        "use server";
-        return x;
-      };
-      /* #__PURE__ */ Object.defineProperty($$hoist_0_f, "name", { value: "f" });
-
-      ;export async function $$hoist_1_h(formData) {
-        "use server";
-        return formData.get(x);
-      };
-      /* #__PURE__ */ Object.defineProperty($$hoist_1_h, "name", { value: "h" });
-
-      ;export function $$hoist_2_w() {
-        "use server";
-      };
-      /* #__PURE__ */ Object.defineProperty($$hoist_2_w, "name", { value: "w" });
-      "
-    `)
-
-    // nothing to encode
-    expect(await testTransform(input, { encode: true })).toBe(
-      await testTransform(input),
-    )
-
-    expect(await testTransformNames(input)).toMatchInlineSnapshot(`
-      [
-        "$$hoist_0_f",
-        "$$hoist_1_h",
-        "$$hoist_2_w",
-      ]
-    `)
-  })
-
-  it('closure', async () => {
-    const input = `
-let count = 0;
-
-function Counter() {
-  const name = "value";
-
-  async function changeCount(formData) {
-    "use server";
-    count += Number(formData.get(name));
-  }
-
-  return "something";
-}
-`
-    expect(await testTransform(input)).toMatchInlineSnapshot(`
-      "
-      let count = 0;
-
-      function Counter() {
-        const name = "value";
-
-        const changeCount = /* #__PURE__ */ $$register($$hoist_0_changeCount, "<id>", "$$hoist_0_changeCount").bind(null, name);
-
-        return "something";
-      }
-
-      ;export async function $$hoist_0_changeCount(name, formData) {
-          "use server";
-          count += Number(formData.get(name));
-        };
-      /* #__PURE__ */ Object.defineProperty($$hoist_0_changeCount, "name", { value: "changeCount" });
-      "
-    `)
-  })
-
-  it('many', async () => {
-    const input = `
-let count = 0;
-
-function Counter() {
-  const name = "value";
-
-  async function changeCount(formData) {
-    "use server";
-    count += Number(formData.get(name));
-  }
-
-  async function changeCount2(formData) {
-    "use server";
-    count += Number(formData.get(name));
-  }
-
-  return "something";
-}
-`
-    expect(await testTransform(input)).toMatchInlineSnapshot(`
-      "
-      let count = 0;
-
-      function Counter() {
-        const name = "value";
-
-        const changeCount = /* #__PURE__ */ $$register($$hoist_0_changeCount, "<id>", "$$hoist_0_changeCount").bind(null, name);
-
-        const changeCount2 = /* #__PURE__ */ $$register($$hoist_1_changeCount2, "<id>", "$$hoist_1_changeCount2").bind(null, name);
-
-        return "something";
-      }
-
-      ;export async function $$hoist_0_changeCount(name, formData) {
-          "use server";
-          count += Number(formData.get(name));
-        };
-      /* #__PURE__ */ Object.defineProperty($$hoist_0_changeCount, "name", { value: "changeCount" });
-
-      ;export async function $$hoist_1_changeCount2(name, formData) {
-          "use server";
-          count += Number(formData.get(name));
-        };
-      /* #__PURE__ */ Object.defineProperty($$hoist_1_changeCount2, "name", { value: "changeCount2" });
-      "
-    `)
-  })
-
-  it('arrow', async () => {
-    const input = `
-let count = 0;
-
-function Counter() {
-  const name = "value";
-
-  return {
-    type: "form",
-    action: (formData) => {
-      "use server";
-      count += Number(formData.get(name));
-    }
-  }
-}
-`
-    expect(await testTransform(input)).toMatchInlineSnapshot(`
-      "
-      let count = 0;
-
-      function Counter() {
-        const name = "value";
-
-        return {
-          type: "form",
-          action: /* #__PURE__ */ $$register($$hoist_0_anonymous_server_function, "<id>", "$$hoist_0_anonymous_server_function").bind(null, name)
-        }
-      }
-
-      ;export function $$hoist_0_anonymous_server_function(name, formData) {
-            "use server";
-            count += Number(formData.get(name));
-          };
-      /* #__PURE__ */ Object.defineProperty($$hoist_0_anonymous_server_function, "name", { value: "anonymous_server_function" });
-      "
-    `)
-
-    expect(await testTransform(input, { encode: true })).toMatchInlineSnapshot(`
-      "
-      let count = 0;
-
-      function Counter() {
-        const name = "value";
-
-        return {
-          type: "form",
-          action: /* #__PURE__ */ $$register($$hoist_0_anonymous_server_function, "<id>", "$$hoist_0_anonymous_server_function").bind(null, __enc([name]))
-        }
-      }
-
-      ;export function $$hoist_0_anonymous_server_function($$hoist_encoded, formData) {
-            const [name] = __dec($$hoist_encoded);
-      "use server";
-            count += Number(formData.get(name));
-          };
-      /* #__PURE__ */ Object.defineProperty($$hoist_0_anonymous_server_function, "name", { value: "anonymous_server_function" });
-      "
-    `)
-  })
-
-  it('higher order', async () => {
-    // packages/react-server/examples/next/app/actions/header/page.tsx
-    // packages/react-server/examples/next/app/actions/header/validator.ts
-    const input = `
-export default function Page() {
-  const x = 0;
-  const action = validator(async (y) => {
-    "use server";
-    return x + y;
-  })
-}
-
-function validator(action) {
-  return async function (arg) {
-    "use server";
-    return action(arg);
-  };
-}
-`
-    expect(await testTransform(input)).toMatchInlineSnapshot(`
-      "
-      export default function Page() {
-        const x = 0;
-        const action = validator(/* #__PURE__ */ $$register($$hoist_0_anonymous_server_function, "<id>", "$$hoist_0_anonymous_server_function").bind(null, x))
-      }
-
-      function validator(action) {
-        return /* #__PURE__ */ $$register($$hoist_1_anonymous_server_function, "<id>", "$$hoist_1_anonymous_server_function").bind(null, action);
-      }
-
-      ;export async function $$hoist_0_anonymous_server_function(x, y) {
-          "use server";
-          return x + y;
-        };
-      /* #__PURE__ */ Object.defineProperty($$hoist_0_anonymous_server_function, "name", { value: "anonymous_server_function" });
-
-      ;export async function $$hoist_1_anonymous_server_function(action, arg) {
-          "use server";
-          return action(arg);
-        };
-      /* #__PURE__ */ Object.defineProperty($$hoist_1_anonymous_server_function, "name", { value: "anonymous_server_function" });
-      "
-    `)
-
-    expect(await testTransform(input, { encode: true })).toMatchInlineSnapshot(`
-      "
-      export default function Page() {
-        const x = 0;
-        const action = validator(/* #__PURE__ */ $$register($$hoist_0_anonymous_server_function, "<id>", "$$hoist_0_anonymous_server_function").bind(null, __enc([x])))
-      }
-
-      function validator(action) {
-        return /* #__PURE__ */ $$register($$hoist_1_anonymous_server_function, "<id>", "$$hoist_1_anonymous_server_function").bind(null, __enc([action]));
-      }
-
-      ;export async function $$hoist_0_anonymous_server_function($$hoist_encoded, y) {
-          const [x] = __dec($$hoist_encoded);
-      "use server";
-          return x + y;
-        };
-      /* #__PURE__ */ Object.defineProperty($$hoist_0_anonymous_server_function, "name", { value: "anonymous_server_function" });
-
-      ;export async function $$hoist_1_anonymous_server_function($$hoist_encoded, arg) {
-          const [action] = __dec($$hoist_encoded);
-      "use server";
-          return action(arg);
-        };
-      /* #__PURE__ */ Object.defineProperty($$hoist_1_anonymous_server_function, "name", { value: "anonymous_server_function" });
-      "
-    `)
-  })
-
-  // edge case found in https://github.com/remix-run/react-router/blob/98367e49900701c460cb08eb16c2441da5007efc/playground/rsc-vite/src/routes/home/home.tsx
-  it('export before import', async () => {
-    const input = `
-export {} from "edge-case";
-import { redirect } from "react-router/rsc";
-
-export default () => {
-  const redirectOnServer = async () => {
-    "use server";
-    throw redirect();
-  };
-}
-`
-    expect(await testTransform(input)).toMatchInlineSnapshot(`
-      "
-      export {} from "edge-case";
-      import { redirect } from "react-router/rsc";
-
-      export default () => {
-        const redirectOnServer = /* #__PURE__ */ $$register($$hoist_0_redirectOnServer, "<id>", "$$hoist_0_redirectOnServer");
-      }
-
-      ;export async function $$hoist_0_redirectOnServer() {
-          "use server";
-          throw redirect();
-        };
-      /* #__PURE__ */ Object.defineProperty($$hoist_0_redirectOnServer, "name", { value: "redirectOnServer" });
-      "
-    `)
   })
 
   it('noExport', async () => {
